code/requirements.txt
pandas
numpy
google-api-python-client
requests
datetime
pyyaml